move various utils to one package
