+++ tests/functional/utils/generic.py
@@ -22,61 +22,28 @@ import requests
 from behave.runner import Context
 from bs4 import BeautifulSoup
 from collections import namedtuple
+from jsonschema import validate
 from langdetect import DetectorFactory, detect_langs
 from random import choice
 from requests import Response
-from requests import Session
-from requests.exceptions import (
-    ChunkedEncodingError,
-    ConnectionError,
-    ConnectTimeout,
-    ContentDecodingError,
-    HTTPError,
-    InvalidHeader,
-    InvalidSchema,
-    InvalidURL,
-    MissingSchema,
-    ProxyError,
-    ReadTimeout,
-    RequestException,
-    RetryError,
-    SSLError,
-    StreamConsumedError,
-    Timeout,
-    TooManyRedirects,
-    UnrewindableBodyError,
-    URLRequired
-)
 from retrying import retry
 from scrapy.selector import Selector
 from termcolor import cprint
-from tests.functional.features.context_utils import (
+
+from tests.functional.schemas.Companies import COMPANIES
+from tests.functional.utils.context_utils import (
     CaseStudy,
     Company,
     Feedback,
     Message
 )
-from tests.functional.features.utils import (
-    Method,
-    assertion_msg,
-    blue,
-    extract_by_css,
-    extract_csrf_middleware_token,
-    green,
-    make_request,
-    rare_word,
-    red,
-    sentence
-)
-from urllib3.exceptions import HTTPError as BaseHTTPError
 
 from tests import get_absolute_url
-from tests.functional.pages import int_api_ch_search
-from tests.functional.steps import PAGE_REGISTRY
 from tests.functional.utils.db_utils import (
     delete_supplier_data,
     get_company_email
 )
+from tests.functional.utils.request import make_request, Method
 from tests.settings import (
     MAILGUN_DIRECTORY_API_USER,
     MAILGUN_DIRECTORY_EVENTS_URL,
@@ -110,16 +77,6 @@ FAS_SUPPORTED_LANGUAGES = {
 }
 
 
-# a list of exceptions that can be thrown by `requests` (and urllib3)
-REQUEST_EXCEPTIONS = (
-    BaseHTTPError, RequestException, HTTPError, ConnectionError, ProxyError,
-    SSLError, Timeout, ConnectTimeout, ReadTimeout, URLRequired,
-    TooManyRedirects, MissingSchema, InvalidSchema, InvalidURL, InvalidHeader,
-    ChunkedEncodingError, ContentDecodingError, StreamConsumedError, RetryError,
-    UnrewindableBodyError
-)
-
-
 def get_file_log_handler(
         log_formatter, log_file=os.path.join(
             ".", "tests", "functional", "reports", "behave.log"),
@@ -150,22 +107,6 @@ def init_loggers(context: Context):
     # Add log file handler to Behave's logging
     context.config.setup_logging(handlers=[log_file_handler])
 
-
-class Method(Enum):
-    """Lists all HTTP methods supported by `requests`."""
-    DELETE = 0
-    GET = 1
-    HEAD = 2
-    OPTIONS = 3
-    PATCH = 4
-    POST = 5
-    PUT = 6
-
-    def __str__(self):
-        return self.name
-
-    def __eq__(self, y):
-        return self.value == y.value
 
 
 def decode_as_utf8(content):
@@ -381,85 +322,29 @@ def log_response(response: Response, *, trim: bool = True):
             logging.debug("RSP Content: %s", content)
 
 
-def make_request(
-        method: Method, url: str, *, session: Session = None,
-        params: dict = None, headers: dict = None, cookies: dict = None,
-        data: dict = None, files: dict = None,
-        allow_redirects: bool = True, auth: tuple = None, trim: bool = True) \
-        -> Response:
-    """Make a desired HTTP request using optional parameters, headers and data.
+def int_api_ch_search(term: str) -> dict:
+    """Will search for companies using provided term.
 
     NOTE:
-    If you want to send a POST/PUT/PATCH request as "multipart/form-data;"
-    rather than a default "application/x-www-form-urlencoded",
-    then provide `data` as `files`.
+    This will validate the response data against appropriate JSON Schema.
 
-    :param method: HTTP method, e.g.: GET, POST, PUT etc
-    :param url: URL that request will be made against
-    :param session: (optional) an instance of requests Session
-    :param params: (optional) query parameters
-    :param headers: (optional) extra request headers. Will not be persisted
-                    across requests, even if using a session.
-    :param cookies: (optional) extra request cookies. Will not be persisted
-                    across requests, even if using a session.
-    :param data: (optional) data to send
-    :param files: (optional) a dict with a file.
-                  For more details please refer to:
-                  http://docs.python-requests.org/en/master/user/quickstart/#post-a-multipart-encoded-file
-    :param allow_redirects: Follow or do not follow redirects
-    :param auth: (optional) authentication tuple, e.g.: ("username", "password")
-    :param trim: (optional) trim long request body/response content if True
-    :return: a response object
+    :param term: search term, can be: company name or number, keywords etc.
+    :type term: str
+    :return: a JSON response from Companies House Search endpoint
     """
-    with assertion_msg("Can't make a request without a valid URL!"):
-        assert url is not None
+    url = get_absolute_url('internal-api:companies-house-search')
+    params = {"term": term}
+    response = make_request(
+        Method.GET, url, params=params, allow_redirects=False)
+    with assertion_msg(
+            "Expected 200 OK from GET %s but instead got {}. In case you're "
+            "getting 301 Redirect then check if you're using correct protocol "
+            "https or http", response.url, response.status_code):
+        assert response.status_code == 200
+    logging.debug("Company House Search result: %s", response.json())
+    validate(response.json(), COMPANIES)
 
-    req = session or requests
-
-    connect_timeout = 3.05
-    read_timeout = 60
-    request_kwargs = dict(
-        url=url, params=params, headers=headers, cookies=cookies, data=data,
-        files=files, allow_redirects=allow_redirects,
-        timeout=(connect_timeout, read_timeout), auth=auth)
-
-    if not allow_redirects:
-        msg = "REQ Follow redirects: disabled"
-        blue(msg)
-        logging.debug(msg)
-
-    try:
-        if method == Method.DELETE:
-            res = req.delete(**request_kwargs)
-        elif method == Method.GET:
-            res = req.get(**request_kwargs)
-        elif method == Method.HEAD:
-            res = req.head(**request_kwargs)
-        elif method == Method.OPTIONS:
-            res = req.options(**request_kwargs)
-        elif method == Method.PATCH:
-            res = req.patch(**request_kwargs)
-        elif method == Method.POST:
-            res = req.post(**request_kwargs)
-        elif method == Method.PUT:
-            res = req.put(**request_kwargs)
-        else:
-            raise KeyError("Unrecognized Method: %s", method.name)
-    except REQUEST_EXCEPTIONS as ex:
-        red("Exception UTC datetime: %s" %
-            datetime.isoformat(datetime.utcnow()))
-        red("{} {}".format(method, url))
-        red("Parameters: {}".format(params))
-        if headers.get('Authorization'):
-            headers['Authorization'] = 'STRIPPED_OUT'
-        red("Headers: {}".format(headers))
-        red("Cookies: {}".format(cookies))
-        red("Data: {}".format(data))
-        red("Files: {}".format(files))
-        raise ex
-
-    log_response(res, trim=trim)
-    return res
+    return response.json()
 
 
 def extract_csrf_middleware_token(response: Response):
@@ -552,64 +437,6 @@ def extract_password_reset_link(payload: str) -> str:
     logging.debug("Found password reset link: %s", password_reset_link)
     return password_reset_link
 
-
-def check_response(response: Response, status_code: int, *,
-                   location: str = None, locations: list = None,
-                   location_starts_with: str = None, body_contains: list = None,
-                   unexpected_strings: list = None):
-    """Check if SUT replied with an expected response.
-
-    :param response: Response object return by `requests`
-    :param status_code: expected status code
-    :param location: (optional) expected value of Location header
-    :param locations: (optional) in one of the Location list
-    :param location_starts_with: (optional) expected leading part of
-                                the Location header
-    :param body_contains: (optional) a list of strings that should be present
-                    in the response content
-    :param unexpected_strings: (optional) a list of strings that should NOT be
-                               present in the response content
-    """
-    with assertion_msg(
-            "Expected %s but got %s", status_code, response.status_code):
-        assert response.status_code == status_code
-
-    if body_contains:
-        with assertion_msg("Expected response with content, got an empty one!"):
-            assert response.content
-        content = response.content.decode("utf-8")
-        for string in body_contains:
-            with assertion_msg("Could not find '%s' in the response", string):
-                assert string in content
-
-    if unexpected_strings:
-        with assertion_msg("Expected response with content, got an empty one!"):
-            assert response.content
-        content = response.content.decode("utf-8")
-        for string in unexpected_strings:
-            with assertion_msg("Found unexpected '%s' in response", string):
-                assert string not in content
-
-    if location:
-        new_location = response.headers.get("Location")
-        with assertion_msg("Expected Location header to be: '%s' but got '%s' "
-                           "instead.", location, new_location):
-            assert new_location == location
-
-    if locations:
-        new_location = response.headers.get("Location")
-        with assertion_msg(
-                "Should redirect to one of these %d locations '%s' but instead "
-                "was redirected to '%s'", len(locations), locations,
-                new_location):
-            assert new_location in locations
-
-    if location_starts_with:
-        new_location = response.headers.get("Location")
-        with assertion_msg(
-                "Expected Location header to start with: '%s' but got '%s' "
-                "instead.", location_starts_with, new_location):
-            assert new_location.startswith(location_starts_with)
 
 
 def get_absolute_path_of_file(filename):
@@ -1037,7 +864,7 @@ def find_active_company_without_fas_profile(alias: str) -> Company:
             logging.debug("Company %s is not registered with FAB: Getting "
                           "it details from CH...", random_company_number)
 
-        json = int_api_ch_search.search(random_company_number)
+        json = int_api_ch_search(random_company_number)
 
         if len(json) == 1:
             exists = True
@@ -1173,7 +1000,7 @@ def update_companies():
             index + 1, companies_number, company.number, company.title))
         blue(progress_message)
         exists, active, is_registered = False, False, False
-        ch_result = int_api_ch_search.search(company.number)
+        ch_result = int_api_ch_search(company.number)
         if len(ch_result) == 1:
             exists = True
             if ch_result[0]["company_status"] == "active":
@@ -1248,6 +1075,7 @@ def get_language_code(language: str):
 
 
 def get_fas_page_url(page_name: str, *, language_code: str = None):
+    from tests.functional.registry import PAGE_REGISTRY
     selector = PAGE_REGISTRY[page_name.lower()]
     url = get_absolute_url(selector)
     if language_code:
@@ -1256,6 +1084,7 @@ def get_fas_page_url(page_name: str, *, language_code: str = None):
 
 
 def get_fabs_page_url(page_name: str, *, language_code: str = None):
+    from tests.functional.registry import PAGE_REGISTRY
     url = get_absolute_url(PAGE_REGISTRY[page_name.lower()])
     if language_code:
         url += "?lang={}".format(language_code)
